Fix setting URL in Firefox
I'm actually not sure why it was broken in Firefox in the first place.
It seemed to add two items to the history, the correct url and "/#",
every time. I played with it bunch and still couldn't figure it out, so
I added this horrible hack to work around the problem. :)