Moving serialization functions into a separate namespace
I've been meaning to do this for a while, but was particularly motivated because
requiring lzw isn't working when building for node.js. This is simpler than
debugging the lzw issue since loading and saving game state isn't on the
critical path for the AI vs AI CLI game runner.